Add delete dialog confirmation

+++ front-end/src/app/menu/arenas/arenas.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { select, Store } from '@ngrx/store'
-import { filter, map, Observable} from 'rxjs';;
+import { filter, map, Observable} from 'rxjs';import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
+;
 import { State } from 'src/app/shared/routing/id-reducer.reducer';
 import { selectId } from 'src/app/shared/routing/id.selectors';
 import { UnsubscribeComponent } from 'src/app/shared/unsubscribe/unsubscribe.component';
@@ -24,7 +26,7 @@ export class ArenasComponent implements OnInit {
   selectedId$ = this.store.pipe(select(selectId));
   arenasList$ : Observable<Arena[]> = new Observable<Arena[]>();
 
-  constructor(private store :Store<State>) { }
+  constructor(private store :Store<State>,private dialog: MatDialog,) { }
 
   ngOnInit(): void { 
     this.store.dispatch(arenasGetAll());
@@ -32,10 +34,23 @@ export class ArenasComponent implements OnInit {
   }
 
   delete(id:number|string|undefined):void{
-    if(!!id){
-      this.store.dispatch(arenaDeleteRow({arenaid:+id}));  
-      this.store.dispatch(arenasGetAll()); 
-    } 
+    this.delete
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent,{
+      data:{
+        message: 'Delete record?',
+        buttonText: {
+          ok: 'Ok',
+          cancel: 'Cancel'
+        }
+      }
+    });
+
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed && !!id) {
+        this.store.dispatch(arenaDeleteRow({arenaid:+id}));  
+        this.store.dispatch(arenasGetAll());         
+      }
+    });
   }
 
 }
